[refactor/#74] Rename shippingDetailsDto to ShippingDetailsDto
